### 2025-11-26 4b79302
CI: Added support for macOS and Ubuntu
