+++ .github/workflows/CI.yml
@@ -19,6 +19,8 @@ jobs:
           - "1.11" # Latest
         os:
           - windows-latest
+          - ubuntu-latest
+          - macos-latest
         arch:
           - x64
     steps:
