### 2025-11-26 2713435
CI: Fix PyCall/Conda mismatch on Ubuntu
