### 2025-11-26 #9114 @mossr
@jbytecode by the way, I fixed the error you were seeing and added Ubuntu and macOS to the CI.

https://github.com/sisl/Crux.jl/actions/runs/19692389677/job/56410811739
