### 2025-12-03 #9114 @DennisWayo
> 👋 @agricolab, @DennisWayo - Would you be willing to assist in reviewing this submission to JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines:
>
> https://joss.readthedocs.io/en/latest/review_criteria.html

Sure.
