### 2025-12-13 #9620 @editorialbot
[REVIEW]: Crux.jl: Deep Reinforcement Learning in Julia
**Submitting author:** @mossr (<a href="http://orcid.org/0000-0003-2403-454X">Robert Moss</a>)
**Repository:** https://github.com/sisl/Crux.jl
**Branch with paper.md** (empty if default branch): master
**Version:** v0.1.4
