### 2025-12-14 ca51524
Update README with links to documentation
