### 2025-12-27 #9620 @DennisWayo
Installation succeeded in a clean temporary Julia environment using Julia 1.12.2 on macOS (Apple Silicon, arm64). Pkg.add("Crux") completed successfully and all dependencies precompiled. GPU-related warnings (cuDNN) were observed, but are expected on CPU-only macOS systems and did not affect functionality. (Obviously, Flux, CUDA, and cuDNN do not support Metal.)

Thanks to @Luis-Varona for pointing out the documentation issue and to @mossr for promptly fixing the broken links. When testing in a clean Julia 1.12 environment, I found that running the Gym/CartPole example still requires installing POMDPGym.jl through a GitHub URL as well as additional Python (Gymnasium) dependencies, which is not immediately obvious from the README or installation documentation. Making this explicit, or providing a fully self-contained minimal example, would further improve reproducibility for new users.

I look forward to seeing this clarified in the documentation and am happy to continue the review once it is addressed.

@jbytecode
