### 2026-01-05 #9620 @Luis-Varona
Hello @jbytecode, just flying back from Christmas vacation, so I should be able to provide a more detailed list of feedback within the next few days.
