### 2026-01-14 #9620 @mossr
Hi @jbytecode, these are great suggestions and I'll work on them this week.
