### 2026-01-19 37e8c85
Added standalone (non-POMDPGym) example, updated docs
