+++ README.md
@@ -1,8 +1,6 @@
 # Crux.jl
 
-[![Documentation](https://img.shields.io/badge/docs-latest-blue.svg)](https://sisl.github.io/Crux.jl/dev/)
-[![Build Status](https://github.com/sisl/Crux.jl/actions/workflows/CI.yml/badge.svg)](https://github.com/sisl/Crux.jl/actions/workflows/CI.yml)
-[![Code Coverage](https://codecov.io/gh/sisl/Crux.jl/branch/master/graph/badge.svg)](https://codecov.io/gh/sisl/Crux.jl)
+[![Documentation](https://img.shields.io/badge/docs-latest-blue.svg)](https://sisl.github.io/Crux.jl/dev/) [![Build Status](https://github.com/sisl/Crux.jl/actions/workflows/CI.yml/badge.svg)](https://github.com/sisl/Crux.jl/actions/workflows/CI.yml) [![Code Coverage](https://codecov.io/gh/sisl/Crux.jl/branch/master/graph/badge.svg)](https://codecov.io/gh/sisl/Crux.jl)
 
 Deep RL library with concise implementations of popular algorithms. Implemented using [Flux.jl](https://github.com/FluxML/Flux.jl) and fits into the [POMDPs.jl](https://github.com/JuliaPOMDP/POMDPs.jl) interface.
 
@@ -40,34 +38,54 @@ Supports CPU and GPU computation and implements deep reinforcement learning, imi
 * [Experience Replay](https://github.com/sisl/Crux.jl/blob/master/src/model_free/cl/experience_replay.jl)
 
 
+## Installation
+
+To install the package, run:
+```julia
+] add Crux
+```
+
 ## Usage
 
-An example usage of the `REINFORCE` algorithm with a simple Flux network for the Cart Pole problem is shown here:
+### Basic Example (Pure Julia)
+
+A minimal example using DQN to solve a GridWorld problem:
+
+```julia
+using Crux
+
+mdp = SimpleGridWorld()
+S = state_space(mdp)
+
+A() = DiscreteNetwork(Chain(Dense(2, 8, relu), Dense(8, 4)), actions(mdp))
+
+solver = DQN(π=A(), S=S, N=100_000)
+policy = solve(solver, mdp)
+```
+
+See the [documentation](https://sisl.github.io/Crux.jl/dev/) for more examples and details.
+
+### Gym Environments
+
+For OpenAI Gym environments like CartPole, install [POMDPGym.jl](https://github.com/ancorso/POMDPGym.jl):
+
+```julia
+] add https://github.com/ancorso/POMDPGym.jl
+```
+
+> **Note:** POMDPGym requires Python with [Gymnasium](https://gymnasium.farama.org/) installed (`pip install gymnasium`).
 
 ```julia
 using Crux, POMDPGym
 
-# Problem setup
 mdp = GymPOMDP(:CartPole)
 as = actions(mdp)
 S = state_space(mdp)
 
-# Flux network: Map states to actions
 A() = DiscreteNetwork(Chain(Dense(dim(S)..., 64, relu), Dense(64, length(as))), as)
 
-# Setup REINFORCE solver
-solver_reinforce = REINFORCE(S=S, π=A())
-
-# Solve the `mdp` to get the `policy`
-policy_reinforce = solve(solver_reinforce, mdp)
-```
-
-
-## Installation
-
-To install the package, run:
-```
-] add Crux
+solver = REINFORCE(S=S, π=A())
+policy = solve(solver, mdp)
 ```
 
 See the [installation documentation](https://github.com/ancorso/POMDPGym.jl?tab=readme-ov-file#installation) for more details on how to install POMDPGym for more environment.
