+++ .github/workflows/CI.yml
@@ -16,7 +16,9 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - "1.11" # Latest
+          - "1.11"
+          - "1.12"
+          - "nightly"
         os:
           - windows-latest
           - ubuntu-latest
