+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Crux"
 uuid = "e51cc422-768a-4345-bb8e-2246287ae729"
 authors = ["Anthony Corso <anthonycorso92@gmail.com>"]
-version = "0.1.3"
+version = "0.1.4"
 
 [deps]
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
