+++ .github/workflows/CI.yml
@@ -18,7 +18,6 @@ jobs:
         version:
           - "1.11"
           - "1.12"
-          - "nightly"
         os:
           - windows-latest
           - ubuntu-latest
