+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"interface/#Library/Interface","page":"Library/Interface","title":"Library/Interface","text":"This section details the interface and functions provided by Crux.jl.","category":"section"},{"location":"interface/#Contents","page":"Library/Interface","title":"Contents","text":"Pages = [\"interface.md\"]","category":"section"},{"location":"interface/#Index","page":"Library/Interface","title":"Index","text":"Pages = [\"interface.md\"]","category":"section"},{"location":"interface/#Reinforcement-Learning","page":"Library/Interface","title":"Reinforcement Learning","text":"RL interfaces are categorized by on-policy and off-policy below.","category":"section"},{"location":"interface/#On-Policy","page":"Library/Interface","title":"On-Policy","text":"","category":"section"},{"location":"interface/#Off-Policy","page":"Library/Interface","title":"Off-Policy","text":"","category":"section"},{"location":"interface/#Imitation-Learning","page":"Library/Interface","title":"Imitation Learning","text":"","category":"section"},{"location":"interface/#Adversarial-RL","page":"Library/Interface","title":"Adversarial RL","text":"","category":"section"},{"location":"interface/#Continual-Learning","page":"Library/Interface","title":"Continual Learning","text":"","category":"section"},{"location":"interface/#Batched-RL","page":"Library/Interface","title":"Batched RL","text":"","category":"section"},{"location":"interface/#Policies","page":"Library/Interface","title":"Policies","text":"","category":"section"},{"location":"interface/#Crux.OnPolicySolver","page":"Library/Interface","title":"Crux.OnPolicySolver","text":"On policy solver type.\n\nFields\n\nagent::PolicyParams Policy parameters (PolicyParams)\nS::AbstractSpace State space\nN::Int = 1000 Number of environment interactions\nΔN::Int = 200 Number of interactions between updates\nmax_steps::Int = 100 Maximum number of steps per episode\nlog::Union{Nothing, LoggerParams} = nothing The logging parameters\ni::Int = 0 The current number of environment interactions\nparam_optimizers::Dict{Any, TrainingParams} = Dict() Training parameters for the parameters\na_opt::TrainingParams Training parameters for the actor\nc_opt::Union{Nothing, TrainingParams} = nothing Training parameters for the critic\n𝒫::NamedTuple = (;) Parameters of the algorithm\ninteraction_storage = nothing If this is initialized to an array then it will store all interactions\npost_sample_callback = (𝒟; kwargs...) -> nothing Callback that that happens after sampling experience\npost_batch_callback = (𝒟; kwargs...) -> nothing Callback that that happens after sampling a batch\n\nOn-policy-specific parameters\n\nλ_gae::Float32 = 0.95 Generalized advantage estimation parameter\nrequired_columns = Symbol[] Extra data columns to store\n\nParameters specific to cost constraints (a separate value network
[…]
