### 2026-01-19 #9620 @mossr
Hi everyone, I've updated the Crux.jl code to `v0.1.4` which includes a standalone example so you don't have to rely on Python/POMDPGym to get going.

```julia
using Crux

mdp = SimpleGridWorld()
S = state_space(mdp)

A() = DiscreteNetwork(Chain(Dense(2, 8, relu), Dense(8, 4)), actions(mdp))

solver = DQN(π=A(), S=S, N=100_000)
policy = solve(solver, mdp)
```

I've included this in the README and the examples in the docs.

And @Luis-Varona, I've added Julia v1.12 to the CI, good call.

Please update via `] up Crux` in pkg mode to get the changes.
