### 2026-03-09 #9620 @jbytecode
@mossr - I've just sent a pull request (https://github.com/sisl/Crux.jl/pull/27). This PR adds citation to Julia which you've already added in your bibtex but omitted in the manuscript.

This paper requires a little bit refactoring. The Summary section has too much citations and a reference to a Figure. The Figure is only a code listing.

- The whole summary part could be moved into the statement of need
- A new and compact summary part could be rewritten, with a small subset of citations and no references to any tables and figures.
- Please just use

backtick backtick backtick julia
println("Hello world")
backtick backtick

to format your Julia code as we do in GitHub markdown pages, so we don't need to use the word Figure as it is just not a figure.

Thank you in advance.
