### 2026-03-19 936c671
Summary refactor based on JOSS (openjournals/joss-reviews/issues/9620) comments.
