+++ joss/paper.md
@@ -26,59 +26,13 @@ header-includes: |
     \usepackage{listings}
     \usepackage{cleveref}
 ---
-\lstdefinelanguage{Julia}{
-    keywords=[3]{solve, actions, state_space, dim, relu, length},
-    keywords=[2]{Crux, POMDPGym, GymPOMDP, DiscreteNetwork, Chain, Dense, ContinuousNetwork, REINFORCE, A2C, PPO, ActorCritic},
-    keywords=[1]{using},
-    sensitive=true,
-    morecomment=[l]{\#},
-    morecomment=[n]{\#=}{=\#},
-    morestring=[s]{"}{"},
-    morestring=[m]{'}{'},
-    alsoletter=!?,
-    literate={,}{{\color[HTML]{0F6FA3},}}1
-             {\{}{{\color[HTML]{0F6FA3}\{}}1
-             {\}}{{\color[HTML]{0F6FA3}\}}}1
-}
-
-\lstset{
-    language         = Julia,
-    backgroundcolor  = \color[HTML]{F2F2F2},
-    basicstyle       = \small\ttfamily\color[HTML]{19177C},
-    numberstyle      = \ttfamily\scriptsize\color[HTML]{7F7F7F},
-    keywordstyle     = [1]{\bfseries\color[HTML]{1BA1EA}},
-    keywordstyle     = [2]{\color[HTML]{0F6FA3}},
-    keywordstyle     = [3]{\color[HTML]{0000FF}},
-    stringstyle      = \color[HTML]{F5615C},
-    commentstyle     = \color[HTML]{AAAAAA},
-    rulecolor        = \color[HTML]{000000},
-    frame=lines,
-    xleftmargin=10pt,
-    framexleftmargin=10pt,
-    framextopmargin=4pt,
-    framexbottommargin=4pt,
-    tabsize=4,
-    captionpos=b,
-    breaklines=true,
-    breakatwhitespace=false,
-    showstringspaces=false,
-    showspaces=false,
-    showtabs=false,
-    columns=fullflexible,
-    keepspaces=true,
-    numbers=none,
-}
-
 
 # Summary
 \href{https://github.com/sisl/Crux.jl}{Crux.jl} is a Julia [@julia] library for deep reinforcement learning (RL) that provides concise, modular implementations of widely used algorithms.
-The package offers CPU/GPU-accelerated training using Flux.jl [@flux] and is built upon shared abstractions (policies, value functions, buffers, objectives, and update rules).
-These abstractions helps with both code reuse and understanding the core differences between algorithms (e.g., their surrogate losses, trust-region constraints, or advantage estimations).
-Crux.jl includes policy-gradient and actor-critic methods such as REINFORCE [@reinforce], PPO [@ppo], and TRPO [@trpo], along with off-policy value-based and actor–critic variants such as DQN [@dqn], TD3 [@td3], and SAC [@sac], with additional support for imitation, offline, adversarial, and continual learning algorithms.
-Shown in \cref{fig:example}, the library integrates with POMDPs.jl [@pomdps_jl] and the Python gymnasium environments [@gym] for reproducible benchmarking and fast experimentation.
+Built on shared abstractions and multiple dispatch, Crux.jl covers policy-gradient, actor-critic, and off-policy methods, with additional support for imitation, offline, adversarial, and continual learning.
+As shown in the following example, the library integrates with POMDPs.jl [@pomdps_jl] and the Python gymnasium environments for reproducible benchmarking and fast experimentation.
 
-\begin{figure}
-\begin{lstlisting}[language=Julia]
+```julia
 using Crux, POMDPGym
 
[…]
