### 2026-03-20 #9620 @mossr
@jbytecode Agreed. I just refactored the Summary section and implemented the `backtick` syntax highlighting instead of the "figure" approach. Also merged your PR to add the Julia citation.

Let me know if there are any other edits you'd like to see.
