### 2026-03-21 #9620 @jbytecode
@mossr - Thank you for the changes. But how about removing the example and putting it in a separate place, for example, `Code Example` ? It's not usual to present code examples in the summary part, are you agree with me?

Meanwhile, we can perform the other things in parallel:

- Please create a tagged release in the software repository. This tagged release is a proof of the reviewed software.
- Use Zenodo to create an archive using this tagged release. In Zenodo, the version, the title, the list of authors and their metadata including their affiliation and ORCID should match those in the submission.
- Report me the DOI, the release (e.g. v1.2.3), the URL of the archive

The archive file is important as it holds the reviewed software so you don't need to update archive when you change something in the manuscript and the bibliography.

Thank you in advance.
