### 2026-03-21 20058ad
JOSS: Moved code block into an Example Usage section
