### 2026-03-21 #9620 @mossr
@jbytecode agreed, I moved the code example to its own section and pushed the changes.

I've also tagged the repo with a v0.1.4 release and created the Zenodo release as well.
- 10.5281/zenodo.19152815
- v0.1.4
- https://doi.org/10.5281/zenodo.19152815
