### 2026-03-21 #9620 @jbytecode
@mossr - The license of software is MIT but the license in Zenodo archive is Creative Commons Attribution 4.0 International. Please fix and ping me again. Thank you in advance.
