### 2026-03-21 #9620 @mossr
@jbytecode Licensed fix on Zenodo to MIT!
