+++ joss/paper.md
@@ -36,8 +36,8 @@ The library integrates with POMDPs.jl [@pomdps_jl] and the Python gymnasium envi
 Reinforcement learning libraries, such as Stable Baselines3 [@sb3] and RLlib [@rllib], often blur the distinction between algorithmic ideas and framework code, hindering fair comparison, reuse, and extension to settings such as partial observability, safety constraints, or offline data.
 In contrast, Crux.jl is a compact, Julia-native framework built on multiple dispatch and Flux.jl [@flux] that factors training into explicit, swappable components---policies, value functions, buffers, objectives, and update rules---offering CPU/GPU-accelerated training.
 These abstractions help with both code reuse and understanding the core differences between algorithms (e.g., their surrogate losses, trust-region constraints, or advantage estimations).
-Crux.jl includes policy-gradient and actor-critic methods such as REINFORCE [@reinforce], PPO [@ppo], and TRPO [@trpo], along with off-policy value-based and actor--critic variants such as DQN [@dqn], TD3 [@td3], and SAC [@sac].
-In contrast to the inheritance-based code for Stable Baseline3, Crux.jl implements the `DQN` solver using a simple `dqn_target` function for the `OffPolicySolver` type, and a separate `td3_target` function for the same `OffPolicySolver` when implementing `TD3`.
+Crux.jl includes policy-gradient and actor-critic methods such as REINFORCE [@reinforce], PPO [@ppo], and TRPO [@trpo], along with off-policy value-based and actor-critic variants such as DQN [@dqn], TD3 [@td3], and SAC [@sac].
+For example, in contrast to the inheritance-based code for Stable Baselines3, Crux.jl implements the `DQN` solver using a simple `dqn_target` function for the `OffPolicySolver` type, and a separate `td3_target` function for the same `OffPolicySolver` when implementing `TD3`.
 This design enables rigorous, reproducible experimentation across RL settings, and integration with POMDPs.jl [@pomdps_jl] and the Python gymnasium environments [@gym] standardizes environment interaction and evaluation.
 In short, Crux.jl provides a principled, composable deep RL framework for Julia that enables rapid ablations, fair baselines, and reproducible results without sacrificing performance or clarity.
 
