### 2026-03-21 #9620 @mossr
@jbytecode I just pushed some minor typo fixes to the paper. Please recreate the proof and I'll give my final proof-read. Thank you!
