### 2026-03-21 #9620 @mossr
@jbytecode Looks good to me!
